Restructured code to help with organization.
