--- src/main/java/amazin/BookRepository.java
@@ -1,10 +1,0 @@
-package amazin;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
-@RepositoryRestResource
-public interface BookRepository extends CrudRepository<Book, Long> {
-    Book findBookById(long id);
-
-}
