+++ src/main/java/amazin/BookStore.java
@@ -1,7 +1,7 @@
 package amazin;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import amazin.model.Book;
+import amazin.repository.BookRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,8 +9,6 @@ import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class BookStore {
-
-    private static final Logger log = LoggerFactory.getLogger(BookStore.class);
 
     public static void main(String[] args) {
         SpringApplication.run(BookStore.class);
