+++ src/main/java/amazin/model/CustomerReview.java
@@ -1,4 +1,4 @@
-package amazin;
+package amazin.model;
 
 import jakarta.persistence.*;
