+++ src/main/java/amazin/repository/BookRepository.java
@@ -1,9 +1,11 @@
-package amazin;
+package amazin.repository;
 
+import amazin.model.Book;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
-    Customer findCustomerById(Integer id);
+public interface BookRepository extends CrudRepository<Book, Long> {
+    Book findBookById(long id);
+
 }
