+++ src/main/java/amazin/repository/CustomerRepository.java
@@ -1,0 +1,10 @@
+package amazin.repository;
+
+import amazin.model.Customer;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+@RepositoryRestResource
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+    Customer findCustomerById(Integer id);
+}
